feat: add event details on hover
